+++ app/main.go
@@ -1,9 +1,9 @@
 package app
 
 import (
+	"github.com/rajihawa/unmask/app/domain"
 	"github.com/rajihawa/unmask/app/repository"
 	usecases "github.com/rajihawa/unmask/app/usecases"
-	"github.com/rajihawa/unmask/domain"
 )
 
 type AppConfig struct {
